backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
google-generativeai==0.8.4
playwright==1.49.0
python-dotenv==1.0.1
pydantic==2.10.4
aiosqlite==0.20.0
